requirements.txt
crewai
python-dotenv
deepseek-sdk
hubspot-api-client
stripe
xero-python
selenium
webdriver-manager
exchangelib
msal
